azure-certifications-study-guide/AI-projects/project-02-customer-service-bot/src/requirements.txt
botbuilder-core==4.14.8
botbuilder-schema==4.14.8
aiohttp==3.9.0
azure-cosmos==4.5.1
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
